+++ tests/formatter/input-output-pairs/parentheses-expressions-removal.in.gd
@@ -7,3 +7,7 @@ class X:
 		var b = {(1):(2)}
 		var c = [("xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"),"x"]
 		var d = {("xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"):("xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx")}
+		var e = x[(1)]
+
+	func bar(x = (1), y := (1)):
+		pass