+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
  - Fixed parent block comments in if-elif-else
+ - Fixed parentheses handling in formatter
 
 ## [3.2.3] 2020-02-04